make util for feed component
