+++ src/pages/Feed/util.ts
@@ -1,4 +1,20 @@
-export default function formatDate(dateString: string): string {
+const getColor = (color: number):string => {
+  if (color >= 0 && color < 100) {
+    return 'rose';
+  } else if (color >= 100 && color < 150) {
+    return 'apricot';
+  } else if (color >= 150 && color < 200) {
+    return 'lemon';
+  } else if (color >= 200 && color < 250) {
+    return 'tea-green';
+  } else if (color >= 250) {
+    return 'jagged-ice';
+  }
+
+  return 'grey-200';
+};
+
+function formatDate(dateString: string): string {
   const date = new Date(dateString);
   const day = date.getDate();
   const month = date.getMonth() + 1;
@@ -13,3 +29,5 @@ export default function formatDate(dateString: string): string {
 
   return `${formattedMonth}/${formattedDay}/${year} ${formattedHours}:${formattedMinutes}`;
 }
+
+export { getColor, formatDate };
